Unblock Pause Menu and adjust shotgun hitbox
pause menu was "blocked" by the new HUD, so now I've moved it into being a child node of the HUD.

shotgun barrel hitbox was very narrow (2 pixels wide) and had a tendency to get stuck between tiles - Excalibur in stone except it's a shotgun.  solution: widened hitbox to outline, now it gets stuck at least less often.  this will need to be a consideration when we start updating the art assets.