experimental mouse aim change
trying out a hybrid of the old aiming (for precision) and a new way using constant_torque.

this so far seems to make clockwise and counter-clockwise turning far more equal.

this uses the dot product of the y transform of the player and the desired aim direction tied to the torque of the player - which happens to be zero if they are at a perfect 90 degree angle - at perfect aim.

however, relying only on this results in a VERY wobbly aim, torquing back and forth.  therefore, at a cutoff point this goes back to using angular velocity to stabilize it when close to where the player wants to aim.